python-bot-server/requirements.txt
python-telegram-bot
asyncio
aiohttp
python-dotenv
cryptography
redis
motor
pymongo
dnspython
bson
logging
typing
datetime
hashlib
hmac
secrets
base64
re
string
random
time
os
json
functools